+++ src/service/layers/application/entry-points/rest/common/index.ts
@@ -1,3 +1,4 @@
 export * from "./handler";
 export * from "./request";
 export * from "./http-methods";
+export * from "./schema";