+++ src/service/layers/infrastructure/orm/adapter-loader.ts
@@ -13,19 +13,19 @@ import type {
  * @param config Mongoose repository configuration
  * @returns ORMPort implementation for Mongoose
  */
-export function createMongooseAdapter<D, I, O>(
+export const createMongooseAdapter = <D, I, O>(
   config: MongooseRepositoryConfig<D>,
-): ORMPort<MongooseRepositoryModel<D>, I, O> {
+): ORMPort<MongooseRepositoryModel<D>, I, O> => {
   return new MongooseAdapter<D, I, O>(config);
-}
+};
 
 /**
  * Creates a Sequelize adapter
  * @param config Sequelize repository configuration
  * @returns ORMPort implementation for Sequelize
  */
-export function createSequelizeAdapter<D extends object, I, O>(
+export const createSequelizeAdapter = <D extends object, I, O>(
   config: SequelizeRepositoryConfig<D>,
-): ORMPort<SequelizeRepositoryModel<D>, I, O> {
+): ORMPort<SequelizeRepositoryModel<D>, I, O> => {
   return new SequelizeAdapter<D, I, O>(config);
-}
+};